// mooring-pin/api-client/models/MarinaModel.ts
export interface MarinaModel {
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    id?: string;
    /**
     * 
     * @type {number}
     * @memberof MarinaModel
     */
    geoJsonId?: number;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    type?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    coordinates?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    canalplanLink?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    website?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    address?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    canalName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MarinaModel
     */
    phoneNumber?: string | null;
    /**
     * 
     * @type {number}
     * @memberof MarinaModel
     */
    distanceFromSearch?: number | null;
    /**
     * 
     * @type {number}
     * @memberof MarinaModel
     */
    distanceFromUser?: number | null;
    /**
     * 
     * @type {MarinaTypeModel}
     * @memberof MarinaModel
     */
    marinaType?: MarinaTypeModel;
    /**
     * 
     * @type {Array<ServiceModel>}
     * @memberof MarinaModel
     */
    services?: Array<ServiceModel> | null;
}